Seperated attachment components into individual files, refactored code to place all chat-related components in own directory for greater clarity
