+++ app/src/main/java/com/example/roomie/components/chat/ChatDataTypes.kt
@@ -1,4 +1,4 @@
-package com.example.roomie.components
+package com.example.roomie.components.chat
 
 import com.google.firebase.Timestamp
