+++ app/src/main/java/com/example/roomie/navigation/Navigation.kt
@@ -3,20 +3,15 @@ package com.example.roomie.navigation
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.navigation.NavType
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import androidx.navigation.navArgument
-import com.example.roomie.components.ChatManager
 import com.example.roomie.components.OnboardingProfileState
 import com.example.roomie.screens.BasicInfoScreen
 import com.example.roomie.screens.ExtraInfoScreen
 import com.example.roomie.screens.MainContentScreen
-import com.example.roomie.screens.ChatsScreen
 import com.example.roomie.screens.ProfileEditorScreen
 import com.example.roomie.screens.ProfileTypeScreen
-import com.example.roomie.screens.SingleChatScreen
 import com.example.roomie.screens.SplashScreen
 import com.google.firebase.Firebase
 import com.google.firebase.auth.auth
