+++ app/src/main/java/com/example/roomie/screens/ChatsScreen.kt
@@ -1,7 +1,7 @@
 package com.example.roomie.screens
 
 import androidx.compose.foundation.background
-import com.example.roomie.components.ChatItem
+import com.example.roomie.components.chat.ChatItem
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.material3.*
@@ -9,8 +9,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavController
 
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.ui.Alignment
 
 import androidx.compose.runtime.*
@@ -18,8 +16,8 @@ import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.ui.graphics.Color
 
-import com.example.roomie.components.ChatManager
-import com.example.roomie.components.Conversation
+import com.example.roomie.components.chat.ChatManager
+import com.example.roomie.components.chat.Conversation
 import com.google.firebase.Firebase
 import com.google.firebase.auth.auth
 import com.google.firebase.firestore.FirebaseFirestore
